Add skeleton of the core-network module
