+++ .editorconfig
@@ -1,0 +1,5 @@
+root = true
+[*]
+# Most of the standard properties are supported
+indent_size=2
+max_line_length=100
