+++ app/src/main/kotlin/io/getstream/avengerschat/view/dm/DirectMessageRepository.kt
@@ -19,10 +19,6 @@ package io.getstream.avengerschat.view.dm
 import androidx.annotation.WorkerThread
 import io.getstream.avengerschat.extensions.EXTRA_TEAM
 import io.getstream.avengerschat.extensions.currentUserId
-import io.getstream.avengerschat.network.Api.STREAM_CHANNEL_TYPE_MESSAGING
-import io.getstream.avengerschat.network.Api.STREAM_USER_ID
-import io.getstream.avengerschat.network.Api.STREAM_USER_ROLE
-import io.getstream.avengerschat.network.Api.STREAM_USER_ROLE_ADMIN
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.api.models.QueryUsersRequest
 import io.getstream.chat.android.client.models.Filters
