+++ app/src/main/res/layout/fragment_live.xml
@@ -32,7 +32,7 @@
             name="adapterStreamDev"
             type="io.getstream.avengerschat.view.live.LiveStreamDevelopersAdapter" />
 
-        <import type="io.getstream.avengerschat.network.Api" />
+        <import type="io.getstream.avengerschat.core.network.Api" />
     </data>
 
     <androidx.core.widget.NestedScrollView
